Bugfixes & Improvements
- Added new translation
- Added validation to "change admin username"
- Added validation to "adding new user (username)"
- Recordings: "Remove all" will now remove the recordings from choosen room (filter)
- Notifications: "Remove all" will now remove the notifications from choosen room (filter)
- Changes Spdy (http/2) to https (http/1) because of callback error
- Rearranged ui/handler init to avoid incoming mqtt/http message before ui finished initialisation
- Fixed a bug where "offline" toast message appeared also for hidden cameras
- Fixed a bug that caused the video processor entered in config.json not to be passed to videoStream
- Fixed a bug that caused that if only one camera is entered in config.json, the camera page was not visible

+++ app/models/recordings.js
@@ -195,13 +195,39 @@ module.exports = (db, camDb, recDb) => {
     
   }
   
-  async function removeAll(){
+  async function removeAll(room){
 
     let recPath = db.get('settings').get('recordings').get('path').value();
     
     try {
     
-      await fs.emptyDir(recPath);
+      if(room.length && !room.includes('all')){
+      
+        debug('Removing all recordings for following rooms: ' + room.toString());
+        
+        let recs = await get();
+        
+        for(const rec of recs){
+          
+          if(room.includes(rec.room)){
+          
+            await fs.remove(recPath + '/' + rec.fileName);
+            
+            if(rec.type === 'Video')
+              await fs.remove(recPath + '/' + rec.fileName.split('.mp4')[0] + '@2.jpeg');
+          
+          }  
+          
+        }
+      
+      }else {
+      
+        if(room.includes('all')){
+          debug('Removing all recordings!');
+          await fs.emptyDir(recPath);
+        }
+        
+      }
       
     } catch(err) {
